% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holiday_zh}
\alias{holiday_zh}
\title{List of Chinese holidays}
\format{
A data frame with 424 rows and 6 variables:
\describe{
  \item{holiday}{Is it a holiday?}
  \item{name}{Name of the holiday}
  \item{wage}{Wage of the holiday}
  \item{date}{Date of the holiday}
  \item{after}{Is it a day after the holiday?}
  \item{target}{Target of the holiday}
}
}
\source{
\url{https://timor.tech/api/holiday/year/<year>}
}
\usage{
holiday_zh
}
\description{
List of Chinese holidays
}
\keyword{datasets}
